Initial version
Initial working version that barely does anything but magically makes ios devices appear for ffmpeg ( when ffmpeg is also using the kCMI lines to enable it to see them )

Note that despite this "observing" connect and "disconnect" neither actually do anything. They don't show output. Additionally that code doesn't even need to be there for this to work. Left it for determination by others why it doesn't actually do anything...